Create README.md
GLOF_Diagnostics_Programs_V1_delta_wing_v1 is an open source Matlab program set for extraction of high-resolution skin friction fields from time sequences of global luminescent oil-film (GLOF) visualization images.  This program set is useful for researchers to obtain relative (normalized) skin friction fields in complex flows.  This folder contains the main program GLOF_Diagnostics_Run, relevant subroutines and sample GLOF images of a 65-deg delta wing. In addition, a paper is included, which describes the principles of the optical flow method, the programs, the selection of the relevant parameters, and examples to demonstrate the applications of this method.